feat(iota-core): remove unused cache implementations
