+++ crates/iota-core/src/execution_cache/unit_tests/notify_read_input_objects_tests.rs
@@ -18,53 +18,8 @@ use iota_types::{
 use tempfile::tempdir;
 use tokio::time::timeout;
 
-use super::{
-    ObjectCacheRead, passthrough_cache::PassthroughCache, writeback_cache::WritebackCache,
-};
+use super::{ObjectCacheRead, writeback_cache::WritebackCache};
 use crate::authority::{AuthorityStore, authority_store_tables::AuthorityPerpetualTables};
-
-/// Trait abstracting cache operations needed for notify_read_input_objects
-/// tests, so the same test logic can run against both WritebackCache and
-/// PassthroughCache.
-trait NotifyReadTestCache: ObjectCacheRead + Send + Sync + 'static {
-    fn write_object_for_testing(&self, object: Object);
-    fn write_marker_for_testing(
-        &self,
-        epoch_id: u64,
-        object_key: &ObjectKey,
-        marker_value: MarkerValue,
-    );
-}
-
-impl NotifyReadTestCache for WritebackCache {
-    fn write_object_for_testing(&self, object: Object) {
-        WritebackCache::write_object_for_testing(self, object);
-    }
-
-    fn write_marker_for_testing(
-        &self,
-        epoch_id: u64,
-        object_key: &ObjectKey,
-        marker_value: MarkerValue,
-    ) {
-        WritebackCache::write_marker_for_testing(self, epoch_id, object_key, marker_value);
-    }
-}
-
-impl NotifyReadTestCache for PassthroughCache {
-    fn write_object_for_testing(&self, object: Object) {
-        PassthroughCache::write_object_for_testing(self, object);
-    }
-
-    fn write_marker_for_testing(
-        &self,
-        epoch_id: u64,
-        object_key: &ObjectKey,
-        marker_value: MarkerValue,
-    ) {
-        PassthroughCache::write_marker_for_testing(self, epoch_id, object_key, marker_value);
-    }
-}
 
 async fn create_store() -> Arc<AuthorityStore> {
     let path = tempdir().unwrap();
@@ -83,17 +38,9 @@ async fn create_writeback_cache() -> Arc<WritebackCache> {
     Arc::new(WritebackCache::new_for_tests(create_store().await))
 }
 
-async fn create_passthrough_cache() -> Arc<PassthroughCache> {
-    let store = create_store().await;
-    let registry = prometheus::Registry::new();
-    Arc::new(PassthroughCache::new_for_tests(store, &registry))
-}
-
-// ---------------------------------------------------------------------------
-// Generic test functions
-// ---------------------------------------------------------------------------
-
-async fn test_immediate_return_canceled_shared_impl(cache: &impl NotifyReadTestCache) {
+#[tokio::test]
+async fn test_writeback_immediate_return_canceled_shared() {
+    let cache = create_writeback_cache().await;
     let canceled_key = InputKey::VersionedObject {
         id: ObjectID::random(),
         version: SequenceNumber::CANCELLED_READ,
@@ -130,7 +77,9 @@ async fn test_immediate_return_canceled_shared_impl(cache: &impl NotifyReadTestC
     assert_eq!(result.len(), 1);
 }
 
-async fn test_immediate_return_cached_object_impl(cache: &impl NotifyReadTestCache) {
+#[tokio::test]
+async fn test_writeback_immediate_return_cached_object() {
+    let cache = create_writeback_cache().await;
     let object_id = ObjectID::random();
     let version = SequenceNumber::from(1);
     let object = Object::with_id_owner_version_for_testing(object_id, version, Owner::Immutable);
@@ -153,7 +102,9 @@ async fn test_immediate_return_cached_object_impl(cache: &impl NotifyReadTestCac
     assert_eq!(result.len(), 1);
 }
 
-async fn test_immediate_return_cached_package_impl(cache: &impl NotifyReadTestCache) {
+#[tokio::test]
+async fn test_writeback_immediate_return_cached_package() {
+    let cache = create_writeback_cache().await;
     let input_keys = vec![InputKey::Package {
         id: IOTA_FRAMEWORK_PACKAGE_ID,
     }];
@@ -169,7 +120,9 @@ async fn test_immediate_return_cached_package_impl(cache: &impl NotifyReadTestCa
     assert_eq!(result.len(), 1);
 }
 
-async fn test_immediate_return_shared_deleted_impl(cache: &impl NotifyReadTestCache) {
+#[tokio::test]
+async fn test_writeback_immediate_return_shared_deleted() {
+    let cache = create_writeback_cache().await;
     let object_id = ObjectID::random();
     let version = SequenceNumber::from(1);
     let epoch_id = 0;
@@ -197,7 +150,9 @@ async fn test_immediate_return_shared_deleted_impl(cache: &impl NotifyReadTestCa
     assert_eq!(result.len(), 1);
 }
 
-async fn test_wait_for_object_impl<C: NotifyReadTestCache>(cache: Arc<C>) {
+#[tokio::test]
+async fn test_writeback_wait_for_object() {
+    let cache = create_writeback_cache().await;
     let object_id = ObjectID::random();
     let version = SequenceNumber::from(1);
 
@@ -257,7 +212,9 @@ async fn test_wait_for_object_impl<C: NotifyReadTestCache>(cache: Arc<C>) {
     assert_eq!(result.len(), 1);
 }
 
-async fn test_wait_for_package_impl<C: NotifyReadTestCache>(cache: Arc<C>) {
+#[tokio::test]
+async fn test_writeback_wait_for_package() {
+    let cache = create_writeback_cache().await;
     let path = Path::new(env!("CARGO_MANIFEST_DIR")).join("../../examples/move/basics");
     let compiled_modules = BuildConfig::new_for_testing()
         .build(&path)
@@ -293,7 +250,9 @@ async fn test_wait_for_package_impl<C: NotifyReadTestCache>(cache: Arc<C>) {
     assert_eq!(result.len(), 1);
 }
 
-async fn test_wait_for_shared_deleted_impl<C: NotifyReadTestCache>(cache: Arc<C>) {
+#[tokio::test]
+async fn test_writeback_wait_for_shared_deleted() {
+    let cache = create_writeback_cache().await;
     let object_id = ObjectID::random();
     let version = SequenceNumber::from(1);
     let epoch_id = 0;
@@ -327,7 +286,9 @@ async fn test_wait_for_shared_deleted_impl<C: NotifyReadTestCache>(cache: Arc<C>
     assert_eq!(result.len(), 1);
 }
 
-async fn test_receiving_object_higher_version_impl(cache: &impl NotifyReadTestCache) {
+#[tokio::test]
+async fn test_writeback_receiving_object_higher_version() {
+    let cache = create_writeback_cache().await;
     let object_id = ObjectID::random();
     let requested_version = SequenceNumber::from(1);
     let higher_version = SequenceNumber::from(2);
@@ -355,108 +316,4 @@ async fn test_receiving_object_higher_version_impl(cache: &impl NotifyReadTestCa
         .unwrap();
 
     assert_eq!(result.len(), 1);
-}
-
-// ---------------------------------------------------------------------------
-// WritebackCache tests
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn test_writeback_immediate_return_canceled_shared() {
-    let cache = create_writeback_cache().await;
-    test_immediate_return_canceled_shared_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_immediate_return_cached_object() {
-    let cache = create_writeback_cache().await;
-    test_immediate_return_cached_object_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_immediate_return_cached_package() {
-    let cache = create_writeback_cache().await;
-    test_immediate_return_cached_package_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_immediate_return_shared_deleted() {
-    let cache = create_writeback_cache().await;
-    test_immediate_return_shared_deleted_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_wait_for_object() {
-    let cache = create_writeback_cache().await;
-    test_wait_for_object_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_wait_for_package() {
-    let cache = create_writeback_cache().await;
-    test_wait_for_package_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_wait_for_shared_deleted() {
-    let cache = create_writeback_cache().await;
-    test_wait_for_shared_deleted_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_writeback_receiving_object_higher_version() {
-    let cache = create_writeback_cache().await;
-    test_receiving_object_higher_version_impl(&*cache).await;
-}
-
-// ---------------------------------------------------------------------------
-// PassthroughCache tests
-// ---------------------------------------------------------------------------
-
-#[tokio::test]
-async fn test_passthrough_immediate_return_canceled_shared() {
-    let cache = create_passthrough_cache().await;
-    test_immediate_return_canceled_shared_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_immediate_return_cached_object() {
-    let cache = create_passthrough_cache().await;
-    test_immediate_return_cached_object_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_immediate_return_cached_package() {
-    let cache = create_passthrough_cache().await;
-    test_immediate_return_cached_package_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_immediate_return_shared_deleted() {
-    let cache = create_passthrough_cache().await;
-    test_immediate_return_shared_deleted_impl(&*cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_wait_for_object() {
-    let cache = create_passthrough_cache().await;
-    test_wait_for_object_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_wait_for_package() {
-    let cache = create_passthrough_cache().await;
-    test_wait_for_package_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_wait_for_shared_deleted() {
-    let cache = create_passthrough_cache().await;
-    test_wait_for_shared_deleted_impl(cache).await;
-}
-
-#[tokio::test]
-async fn test_passthrough_receiving_object_higher_version() {
-    let cache = create_passthrough_cache().await;
-    test_receiving_object_higher_version_impl(&*cache).await;
 }
