+++ crates/iota-node/src/lib.rs
@@ -409,7 +409,6 @@ impl IotaNode {
 
         let cache_traits = build_execution_cache(
             &config.execution_cache_config,
-            &epoch_start_configuration,
             &prometheus_registry,
             &store,
             backpressure_manager.clone(),
