+++ crates/iota-swarm/src/memory/swarm.rs
@@ -13,7 +13,7 @@ use std::{
 use anyhow::Result;
 use futures::future::try_join_all;
 use iota_config::{
-    ExecutionCacheConfig, ExecutionCacheType, IOTA_GENESIS_FILENAME, NodeConfig,
+    ExecutionCacheConfig, IOTA_GENESIS_FILENAME, NodeConfig,
     node::{AuthorityOverloadConfig, DBCheckpointConfig, GrpcApiConfig, RunWithRange},
     p2p::DiscoveryConfig,
 };
@@ -61,7 +61,6 @@ pub struct SwarmBuilder<R = OsRng> {
     db_checkpoint_config: DBCheckpointConfig,
     num_unpruned_validators: Option<usize>,
     authority_overload_config: Option<AuthorityOverloadConfig>,
-    execution_cache_type: Option<ExecutionCacheType>,
     execution_cache_config: Option<ExecutionCacheConfig>,
     data_ingestion_dir: Option<PathBuf>,
     fullnode_run_with_range: Option<RunWithRange>,
@@ -97,7 +96,6 @@ impl SwarmBuilder {
             db_checkpoint_config: DBCheckpointConfig::default(),
             num_unpruned_validators: None,
             authority_overload_config: None,
-            execution_cache_type: None,
             execution_cache_config: None,
             data_ingestion_dir: None,
             fullnode_run_with_range: None,
@@ -135,7 +133,6 @@ impl<R> SwarmBuilder<R> {
             db_checkpoint_config: self.db_checkpoint_config,
             num_unpruned_validators: self.num_unpruned_validators,
             authority_overload_config: self.authority_overload_config,
-            execution_cache_type: self.execution_cache_type,
             execution_cache_config: self.execution_cache_config,
             data_ingestion_dir: self.data_ingestion_dir,
             fullnode_run_with_range: self.fullnode_run_with_range,
@@ -295,12 +292,6 @@ impl<R> SwarmBuilder<R> {
         self
     }
 
-    pub fn with_execution_cache_type(mut self, execution_cache_type: ExecutionCacheType) -> Self {
-        assert!(self.execution_cache_type.is_none());
-        self.execution_cache_type = Some(execution_cache_type);
-        self
-    }
-
     pub fn with_execution_cache_config(
         mut self,
         execution_cache_config: ExecutionCacheConfig,
@@ -411,10 +402,6 @@ impl<R: rand::RngCore + rand::CryptoRng> SwarmBuilder<R> {
             if let Some(authority_overload_config) = self.authority_overload_config {
                 config_builder =
                     config_builder.with_authority_overload_config(authority_overload_config);
-            }
-
-            if let Some(execution_cache_type) = self.execution_cache_type {
-                config_builder = config_builder.with_execution_cache_type(execution_cache_type);
             }
 
             if let Some(execution_cache_config) = self.execution_cache_config {
