+++ crates/test-cluster/src/lib.rs
@@ -14,8 +14,8 @@ use std::{
 use futures::future::join_all;
 use iota_common::fatal;
 use iota_config::{
-    Config, ExecutionCacheConfig, ExecutionCacheType, IOTA_CLIENT_CONFIG, IOTA_KEYSTORE_FILENAME,
-    IOTA_NETWORK_CONFIG, NodeConfig, PersistedConfig,
+    Config, ExecutionCacheConfig, IOTA_CLIENT_CONFIG, IOTA_KEYSTORE_FILENAME, IOTA_NETWORK_CONFIG,
+    NodeConfig, PersistedConfig,
     genesis::Genesis,
     node::{AuthorityOverloadConfig, DBCheckpointConfig, GrpcApiConfig, RunWithRange},
 };
@@ -989,7 +989,6 @@ pub struct TestClusterBuilder {
     num_unpruned_validators: Option<usize>,
     config_dir: Option<PathBuf>,
     authority_overload_config: Option<AuthorityOverloadConfig>,
-    execution_cache_type: Option<ExecutionCacheType>,
     execution_cache_config: Option<ExecutionCacheConfig>,
     data_ingestion_dir: Option<PathBuf>,
     fullnode_run_with_range: Option<RunWithRange>,
@@ -1023,7 +1022,6 @@ impl TestClusterBuilder {
             num_unpruned_validators: None,
             config_dir: None,
             authority_overload_config: None,
-            execution_cache_type: None,
             execution_cache_config: None,
             data_ingestion_dir: None,
             fullnode_run_with_range: None,
@@ -1227,12 +1225,6 @@ impl TestClusterBuilder {
         self
     }
 
-    pub fn with_execution_cache_type(mut self, config: ExecutionCacheType) -> Self {
-        assert!(self.network_config.is_none());
-        self.execution_cache_type = Some(config);
-        self
-    }
-
     pub fn with_execution_cache_config(mut self, config: ExecutionCacheConfig) -> Self {
         assert!(self.network_config.is_none());
         self.execution_cache_config = Some(config);
@@ -1358,10 +1350,6 @@ impl TestClusterBuilder {
 
         if let Some(authority_overload_config) = self.authority_overload_config.take() {
             builder = builder.with_authority_overload_config(authority_overload_config);
-        }
-
-        if let Some(execution_cache_type) = self.execution_cache_type.take() {
-            builder = builder.with_execution_cache_type(execution_cache_type);
         }
 
         if let Some(execution_cache_config) = self.execution_cache_config.take() {
